requirements.txt
notion-client
openai
tiktoken
python-dotenv
flask